requirements.txt
antlr4-python3-runtime==4.7.1
Django==2.1.2
django-cors-headers==2.4.0
django-js-asset==1.1.0
djangorestframework==3.8.2
lazy-object-proxy==1.3.1
psycopg2==2.7.6.1
pytz==2018.5
